Add files via upload
The goal of this project was to determine whether Risk-Screening Environmental Indicators could significantly predict return from various industries measured using various ETFs. The data was modified from the EPAs website to enable a multi-variate regression. The dependent variable was the returns from various ETFs that represented each industry from the EPAs data on the environmental impact from each industry. The results were surprising in multiple ways. Ultimately, I concluded that the environmental impact from any industry in the United States does not impact financial performance of the companies in that sector, which is consistent with the premise that current tax incentives and environmental impact incentives are not effectively aligned at a federal level.